feat: rename general/enter,exit paths to guests/check-in,check-out
